fix(bug): disallow deletion of authors with courses
- add check to disallow delete of authos with courses

+++ src/components/author/AuthorsPage.js
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import { Link } from 'react-router-dom';
+import toastr from 'toastr';
 import * as authorActions from '../../actions/authorActions';
 import { sortAuthorsByName } from '../../utils/helpers';
 import AuthorList from './AuthorList';
@@ -16,6 +17,9 @@ class AuthorsPage extends React.Component {
   }
 
   deleteAuthor(author) {
+    if (this.props.courses.filter(course => course.authorId === author.id).length > 0) {
+      return toastr.error('Author with courses cannot be deleted!');
+    }
     this.props.actions.deleteAuthor(author);
   }
 
@@ -33,7 +37,6 @@ class AuthorsPage extends React.Component {
         {authors.length > 0 
           && <AuthorList authors={authors} onDeleteAuthor={this.deleteAuthor} />
         }
-        
       </div>
     );
   }
@@ -41,11 +44,13 @@ class AuthorsPage extends React.Component {
 
 AuthorsPage.propTypes = {
   authors: PropTypes.array.isRequired,
+  courses: PropTypes.array.isRequired,
   actions: PropTypes.object.isRequired
 };
 
 const mapStateToProps = (state, ownProps) => ({
-  authors: sortAuthorsByName([...state.authors]) // So as not to mutate the redux state
+  authors: sortAuthorsByName([...state.authors]),
+  courses: state.courses // So as not to mutate the redux state
 });
 
 const mapDispatchToProps = dispatch => ({
